% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adstockDeriveAppend.R
\name{adstock.DeriveAppend}
\alias{adstock.DeriveAppend}
\title{Finds the optimum Adstock rate and applies it, appending the original table.}
\usage{
adstock.DeriveAppend(data, y, x, lim = FALSE)
}
\arguments{
\item{data}{data.frame that holds the data (independent and dependent)}

\item{y}{Dependent variable e.g. Sales}

\item{x}{Independent variable, of which we want to derive optimum adstock. Can be a vector of variables}

\item{lim}{Numeric. Upper limit for adstock rate}
}
\description{
Derives best adstock rate, based on lowest SSE, and adds the the original data.table.
}
\keyword{adstock}
\keyword{append}
\keyword{derive}

