% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adstock.maxDecay.R
\name{adstock.maxDecay}
\alias{adstock.maxDecay}
\title{Adstocks a variable - max decay.}
\usage{
adstock.maxDecay(data, variable, adstock.rates, max.decay)
}
\arguments{
\item{data}{data.table that holds the variable to be transformed}

\item{variable}{Character string of the variable to be transformed}

\item{adstock.rates}{One (or many) rates to adstock the variable}

\item{max.decay}{Number of periods that the adstock takes to decay to 0}
}
\description{
User sets adstock rates, and period where adstock decays to 0.
}
\examples{
x <- data.table::fread(system.file("extdata", "TVRs.csv", package = "nladwa"))
x[, Week := excelDateToRDate(Week)]

out1 <- adstock.maxDecay(data = x, variable = "TVRs", adstock.rates = 0.5, max.decay = 2)
out2 <- adstock.maxDecay(data = x, variable = "TVRs", adstock.rates = 0.5, max.decay = 10)

data.table::setkey(out1, Week)
data.table::setkey(out2, Week)

out <- out1[out2]
out[, i.TVRs := NULL]

rm(out1, out2)

plot.out <- data.table::melt(out, id.vars = "Week")

ggplot2::ggplot(plot.out, ggplot2::aes(x = Week, y = value, colour = variable)) +
  ggplot2::geom_line()
}
\keyword{adstock}
\keyword{decay}
