% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adstock.R
\name{adstock}
\alias{adstock}
\title{Simple Adstock.}
\usage{
adstock(x, adstock.rate)
}
\arguments{
\item{adstock.rate}{Rate to adstock the variable, between 0 and 1.}

\item{data}{data.table that holds the variable to be transformed}

\item{variable}{Character string of the variable to be transformed}
}
\description{
Adstocks one variable, returns a numeric vector.
}
\keyword{adstock}
\keyword{simple}

