% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adstockSimple.R
\name{adstockSimple}
\alias{adstockSimple}
\title{Simple Adstock.}
\usage{
adstockSimple(x, adstock.rate = 0)
}
\arguments{
\item{adstock.rate}{Rate to adstock the variable}

\item{data}{data.table that holds the variable to be transformed}

\item{variable}{Character string of the variable to be transformed}
}
\description{
Adstocks one variable, returns a numeric vector.
}
\keyword{adstock}
\keyword{simple}

