% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asFormula.R
\name{asFormula}
\alias{asFormula}
\title{Columns names to a formula layout.}
\usage{
asFormula(x)
}
\arguments{
\item{x}{vector or data.table or data.frame to convert to formula}
}
\description{
Outputs columns of a t
}
\examples{
asFormula(data.frame(mtcars))

asFormula(data.table(mtcars))

asFormula(letters)
}
\keyword{formula}
\keyword{lm}
\keyword{names}
