% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATAN.R
\name{ATAN}
\alias{ATAN}
\title{ATAN transformation}
\usage{
ATAN(data, variable, rate)
}
\arguments{
\item{data}{data.table that holds the variable to be transformed}

\item{variable}{Character string of the variable to be transformed}

\item{rate}{One (or many) rates to adstock the variable}
}
\description{
User sets rates, which transforms the input variable, outputting in a new table.
}
\keyword{ATAN}

