% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsToNumeric.R
\name{colsToNumeric}
\alias{colsToNumeric}
\title{Converts selected columns to numeric}
\usage{
colsToNumeric(x, colsToTransform, overwrite = FALSE)
}
\arguments{
\item{x}{data.table.}

\item{colsToTransform}{Character. Vector of character columns to convert.}

\item{overwrite}{Logical. Default = FALSE, and just returns the transformed columns. TRUE = overwrites the original columns.}
}
\description{
Removes all punctuation, except "-" and ".", and converts to numeric. Also divides character with "%" by 100 automatically.
}
\examples{
mtcars.fudged <- data.table(mtcars)
mtcars.fudged[, `:=`(mpg = paste0(mpg, "\%"),
                     qsec = gsub("\\\\.", ",", qsec),
                     drat = paste0("-", drat))]

colsToTransform <- names(Filter(is.character, mtcars.fudged))

colsToNumeric(mtcars.fudged, colsToTransform)

str(mtcars.fudged)

colsToNumeric(mtcars.fudged, colsToTransform, overwrite = TRUE)

str(mtcars.fudged)
}
\keyword{columns}
\keyword{sum}
