% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corPlot.R
\name{corPlot}
\alias{corPlot}
\title{Plot correlation matrix.}
\usage{
corPlot(x)
}
\arguments{
\item{x}{data.table}
}
\description{
Prints a ggplot of selected variables over time.
}
\examples{
corPlot(mtcars)
}
\keyword{correlation}
\keyword{plot}

