% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutVar.R
\name{cutVar}
\alias{cutVar}
\title{Cuts numeric variable into groups}
\usage{
cutVar(x, breaks, midpoints, labels = NULL, dp = 2, ...)
}
\arguments{
\item{x}{Vector of numeric.}

\item{breaks}{Integer. Number of groups to create.}

\item{midpoints}{Logical. Return midpoint of groups, rather than character?}

\item{dp}{Integer. Number of decimal places if midpoints == TRUE.}
}
\description{
Builds on base cut function
}
\examples{
vec <- c(-10, mtcars$mpg)

cutVar(vec, breaks = 3)

cutVar(vec, breaks = 3, midpoints = TRUE)
}
\keyword{cut}
