% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelDateToRDate.R
\name{excelDateToRDate}
\alias{excelDateToRDate}
\title{Excel date to R date class}
\usage{
excelDateToRDate(Date)
}
\arguments{
\item{Date}{Character. In dd/mm/yyyy format.}
}
\description{
dd/mm/yyyy to yyyy-mm-dd
}
\examples{
x <- "31/01/2015"
print(x); class(x)
x <- excelDateToRDate(x)
print(x); class(x)
}
\keyword{date}
\keyword{excel}

