% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findCharAcrossCols.R
\name{findCharAcrossCols}
\alias{findCharAcrossCols}
\title{Find characters across all columns}
\usage{
findCharAcrossCols(x, string)
}
\arguments{
\item{x}{A data.table or data.frame}

\item{string}{A character string that you want to find in x}
}
\description{
Searches each column for a particular string
}
\examples{
findCharAcrossCols()
}
\keyword{character}
\keyword{unique}

