% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fixNames.R
\name{fixNames}
\alias{fixNames}
\title{Fixes column names}
\usage{
fixNames(x)
}
\arguments{
\item{x}{data.table. With weird column names.}
}
\description{
Replaces and weird characters or spaces with dots
}
\examples{
fixNames()
}
\keyword{names}

