% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freadExcel.R
\name{freadExcel}
\alias{freadExcel}
\title{Reads Excel table from file to data.table}
\usage{
freadExcel(file, date.index = NULL, sheet = 1, startRow = 1,
  origin = "1900-01-01")
}
\arguments{
\item{file}{Character. Path to file.}

\item{date.index}{Character. Optional, column name of date variable.}

\item{sheet}{Integer. Sheet number of table (from open.xlsx)}

\item{startRow}{Integer. Row that the data starts (from open.xlsx)}

\item{origin}{Character. from open.xlsx}
}
\description{
Create data.table from Excel table that's in an XLSX file.
}
\keyword{excel}
\keyword{fread}
\keyword{xlsx}

