% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freadExcelClipboard.R
\name{freadExcelClipboard}
\alias{freadExcelClipboard}
\title{Reads Excel table from clipboard to data.table}
\usage{
freadExcelClipboard(...)
}
\arguments{
\item{No}{arguments}
}
\description{
Create data.table from Excel table that's held in the clipboard.
}
\keyword{excel}
