% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.contribution.thresholds.R
\name{gam.contribtion.thresholds}
\alias{gam.contribtion.thresholds}
\title{Outputs point where contribution is closest to 0, and max point}
\usage{
gam.contribtion.thresholds(gam.contribution.object)
}
\arguments{
\item{gam.contribution.object}{gam.contribution object.}
}
\description{
Takes a gam.contribution object.
}
\keyword{GAM}
\keyword{contribution}
\keyword{threshold}

