% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.contribution.ggplot.R
\name{gam.contribution.ggplot}
\alias{gam.contribution.ggplot}
\title{Plots the contribution of a GAM's smoothed independent variables}
\usage{
gam.contribution.ggplot(gam.object, incl.intercept = TRUE,
  scales = "free_x")
}
\arguments{
\item{gam.object}{GAM object. A GAM saved to a variable.}
}
\description{
Takes GAM and visualises it.
}
\examples{
# see example from gam.contribution
}
\keyword{contribution}
\keyword{gam}
\keyword{ggplot}

