% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_shade.R
\name{geom_shade}
\alias{geom_shade}
\title{Geom for shading graph to highlight periods.}
\usage{
geom_shade(data, variable, date.index, alpha = 0.2, overlap = 3.5,
  fill = "pink")
}
\arguments{
\item{data}{data.table that holds the data}

\item{date.index}{Character. Date index of x, e.g. "Week", need to be of Date class.}

\item{variable.}{Numeric. Dummy variable, 1 = plot.}

\item{alpha.}{Transparency of bar shades. 1 = opaque, 0 = transparent.}

\item{overlap.}{Numeric. Overlap of the bars, seems to be OK left as 3.5.}

\item{fill.}{Character. Colour for the bars.}
}
\description{
Plots "recession bars" on a graph. Note - only really good for plotting one variable.
}
\examples{
x <- data.table::fread(system.file("extdata", "lm.csv", package = "nladwa"))
x <- x[complete.cases(x)]
x[, Week := as.Date(Week)]
x[, dummy := 0]
x[sample(1:nrow(x), 60), dummy := 1]

ggplot() +
  geom_shade(data = x, variable = "dummy", date.index = "Week", alpha = 0.5) +
  geom_line(data = x, aes(x = Week, y = TotalGrossLeads))

warning("This function is only really any good for plotting one variable")
}
\keyword{ggplot}
\keyword{shade}

