% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMonth.R
\name{getMonth}
\alias{getMonth}
\title{Gets month from a date}
\usage{
getMonth(Date)
}
\arguments{
\item{Date}{Date class object.}
}
\description{
Gets month from a date, has to be Date class
}
\examples{
x <- as.Date("2015-01-12")
getMonth(x)
}
\keyword{month}
