% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getWeek.R
\name{getWeek}
\alias{getWeek}
\title{Gets week from a date}
\usage{
getWeek(Date, order = "Mon-Sun")
}
\arguments{
\item{Date}{Date class object}

\item{order}{"Mon-Sun" or "Sun-Mon"}
}
\description{
Gets week from a date, has to be Date class
}
\examples{
getWeek()
}
\keyword{week}

