% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeekInMonth.R
\name{getWeekInMonth}
\alias{getWeekInMonth}
\title{Gets week in month from a date}
\usage{
getWeekInMonth(Week)
}
\arguments{
\item{Week}{Date class object.}
}
\description{
Gets week number (in month) from a week, has to be Date class
}
\keyword{month}
\keyword{number}
\keyword{week}

