% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotToPPT.R
\name{ggplotToPPT}
\alias{ggplotToPPT}
\title{Exports a graph for presentations}
\usage{
ggplotToPPT(filename = NULL, plot, units = "cm", width = 43.2,
  height = 13.7, font.size = 17)
}
\arguments{
\item{filename}{Character. Filename of image to be created.}

\item{plot}{Character. Name of ggplot object.}

\item{units}{Character. Unit of measurement for dimensions; "in", "cm", or "mm". See "ggsave" documentation.}
}
\description{
Exports ggplot graph to PNG, readable in PPT presentations.
}
\keyword{PPT}
\keyword{ggplot}

