% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagVariable.R
\name{lagVariable}
\alias{lagVariable}
\title{Lags a variable by x observations}
\usage{
lagVariable(variable, lag.amount)
}
\arguments{
\item{variable}{A vector, or column of a data.frame/table}

\item{lag.amount}{Periods to lag the variable by.}
}
\description{
Lags a variable, adding NAs at the front.
}
\examples{
print(1:10); print(lag.variable(variable = 1:10, lag.amount = 3))
}
\keyword{lag}

