% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.actualvfitted.ggplot.R
\name{lm.actualvfitted.ggplot}
\alias{lm.actualvfitted.ggplot}
\title{Plots actual v fitted}
\usage{
lm.actualvfitted.ggplot(lm.actualvfitted.object)
}
\arguments{
\item{lm.contribution.object}{lm.contribution object. Output of lm.contribtion saved as an object.}
}
\description{
Takes a lm.actualvfitted object and plots.
}
\keyword{actual}
\keyword{fitted}
\keyword{ggplot}
\keyword{lm}

