% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.actualvfitted.R
\name{lm.actualvfitted}
\alias{lm.actualvfitted}
\title{Create a data.table for actual v fitted}
\usage{
lm.actualvfitted(lm.contribution.object)
}
\arguments{
\item{lm.contribution.object}{lm.contribution object. Output of lm.contribtion saved as an object.}
}
\description{
Takes a lm.contribution object and outputs an actual vs fitted table.
}
\keyword{actual}
\keyword{fitted}
\keyword{lm}
