% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lm.contribution.ggplot.R
\name{lm.contribution.ggplot}
\alias{lm.contribution.ggplot}
\title{Plots the contribution of a lm's independent variables}
\usage{
lm.contribution.ggplot(lm.contribution.object, date.index = NULL,
  dependent.var = NULL)
}
\arguments{
\item{date.index}{Character. Only define if error. The date index of the lm(data = ?) variable, as a character, e.g. "Week".}

\item{dependent.var}{Character. Only define if error. The dependent variable of the lm.}

\item{lm.contibution.object}{The output of lm.contribution.}
}
\description{
Takes lm.contribution output and visualises it.
}
\examples{
lm.contribution.ggplot()
}
\keyword{contribution}
\keyword{ggplot}
\keyword{lm}

