% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lm.contribution.R
\name{lm.contribution}
\alias{lm.contribution}
\title{Outputs the contribution of each lm independent variable}
\usage{
lm.contribution(lm.object, date.index)
}
\arguments{
\item{lm.object}{A lm() saved to a variable}

\item{date.index}{The date index of the lm(data = ?) variable, as a character, e.g. "Week"}
}
\description{
Takes a lm and applies coefficients to show how each independent variable contributes to the dependent variable. IMPORTANT: lm(data = ?) has to be a data.table.
}
\examples{
lm.contribution()
}
\keyword{contribution}
\keyword{lm}

