% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.contribution.summary.R
\name{lm.contribution.summary}
\alias{lm.contribution.summary}
\title{Outputs the contribution of each lm independent variable - at a top level}
\usage{
lm.contribution.summary(lm.obj, date.index)
}
\arguments{
\item{date.index}{Character. The date index of the lm(data = ?) variable.}

\item{lm.object}{lm object. A lm saved to a variable.}
}
\description{
Takes a lm and applies coefficients to show how each independent variable contributes to the dependent variable, but grouped by whatever is before the first "." in the independent variable. "Ext." and the intercept are grouped to "Base". IMPORTANT: lm(data = ?) has to be a data.table.
}
\keyword{contribution}
\keyword{lm}
\keyword{summary}
