% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.decomp.R
\name{lm.decomp}
\alias{lm.decomp}
\title{Sums the decomp of a linear model.}
\usage{
lm.decomp(lm.contribution.object)
}
\arguments{
\item{lm.contibution.object}{The output of lm.contribution.}
}
\description{
Takes lm.contribution output and sums all the independent variables' contribution to the KPI.
}
\keyword{contribution}
\keyword{decomp}
\keyword{lm}

