% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.fitconf.ggplot.R
\name{lm.fitconf.ggplot}
\alias{lm.fitconf.ggplot}
\title{Plot of lm.fitconf.}
\usage{
lm.fitconf.ggplot(lm.fitconf.object)
}
\arguments{
\item{lm.fitconf.object}{Output of lm.fitconf function.}
}
\description{
Outputs a ggplot of fit and confidence interval, over time.
}
\keyword{confidence}
\keyword{fit}
\keyword{lm}
\keyword{plot}

