% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.fitconf.R
\name{lm.fitconf}
\alias{lm.fitconf}
\title{Fit of a lm with confidence intervals.}
\usage{
lm.fitconf(lm.object, date.index, conf.level = 0.95)
}
\arguments{
\item{lm.object}{Output of lm function.}

\item{date.index}{Character. The column from the model source data that is the date variable.}

\item{conf.level}{Confidence level, default = 95%.}
}
\description{
data.table with date index, model fit, and lower and upper intervals.
}
\keyword{confidence}
\keyword{fit}
\keyword{lm}

