% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.nladwa.R
\name{lm.nladwa}
\alias{lm.nladwa}
\title{Creates list of all functions associated with lm from this package.}
\usage{
lm.nladwa(x, formula, ..., date.index = "Week", conf.level = 0.95,
  test.dep.var = FALSE)
}
\arguments{
\item{x}{data.table with all the data for modelling.}

\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{conf.level}{confidence level, default = 95\%.}

\item{data.index}{Character. The date index of the lm(data = ?) variable.}

\item{test.dep.varCharacter.}{Defaults to FALSE which replaces ".Train" in the dependent variable (from lm.object), with ".Train". If you know that variable doesn't exist, enter the variable name here.}
}
\description{
Take a lm object, and some arguments to streamline modelling.
}
\keyword{lm}
