% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.trainAndTest.ggplot.R
\name{lm.trainAndTest.ggplot}
\alias{lm.trainAndTest.ggplot}
\title{Create a data.table for actual v fitted, with train and test sets}
\usage{
lm.trainAndTest.ggplot(lm.trainAndTest.object, date.index = "Week")
}
\arguments{
\item{lm.trainAndTest.object}{The output of lm.trainAndTest.}

\item{date.index}{Character. Only define if error. The date index of the lm(data = ?) variable, as a character, e.g. "Week".}
}
\description{
Takes a lm object and outputs an actual vs fitted table.
}
\examples{
x <- data.table::fread(system.file("extdata", "lm.csv", package = "nladwa"))
x[, Week := excelDateToRDate(Week)]
lm1 <- lm(TotalGrossLeads.Train ~ Search.Spend + Competitor.Spend, data = x)
lm1.trainAndTest <- lm.trainAndTest(lm.object = lm1, date.index = "Week")
print(lm1.trainAndTest)

lm.trainAndTest.ggplot(lm1.trainAndTest)

lm1.trainAndTest[complete.cases(lm1.trainAndTest) & Type == "Test", MAPE(Actual = Actual, Fitted = Fitted)]
}
\keyword{actual}
\keyword{fitted}
\keyword{ggplot}
\keyword{lm}
\keyword{test}
\keyword{train}
