% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.trainAndTest.R
\name{lm.trainAndTest}
\alias{lm.trainAndTest}
\title{Create a data.table for actual v fitted, with train and test sets}
\usage{
lm.trainAndTest(lm.object, date.index = "Week", test.dep.var = FALSE)
}
\arguments{
\item{lm.object}{lm object. A lm saved to a variable.}

\item{date.index}{Character. The date index of the lm(data = ?) variable.}

\item{test.dep.var}{Character. Defaults to FALSE which replaces ".Train" in the dependent variable (from lm.object), with ".Train". If you know that variable doesn't exist, enter the variable name here.}
}
\description{
Takes a lm object and outputs an actual vs fitted table.
}
\examples{
x <- data.table::fread(system.file("extdata", "lm.csv", package = "nladwa"))
x[, Week := excelDateToRDate(Week)]

x <- TrainAndTest.break(x, "TotalGrossLeads", to.break = "2015-04-06", append = TRUE)
# 06/04/2015

lm1 <- lm(TotalGrossLeads.Train ~ Search.Spend + Competitor.Spend, data = x)
lm1.trainAndTest <- lm.trainAndTest(lm.object = lm1, date.index = "Week")
print(lm1.trainAndTest)

lm.trainAndTest.ggplot(lm1.trainAndTest)

MAPE.trainAndTest(lm1.trainAndTest)
}
\keyword{actual}
\keyword{fitted}
\keyword{lm}
\keyword{test}
\keyword{train}
