% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmToCSV.R
\name{lmToCSV}
\alias{lmToCSV}
\title{Linear model summary to CSV}
\usage{
lmToCSV(lm.object, filename)
}
\arguments{
\item{lm.object}{lm object. A lm saved to a variable.}

\item{filename}{Character string. What the output should be saved as.}
}
\description{
Writes summary(lm.object) to a CSV.
}
\keyword{CSV}
\keyword{lm}

