% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPE.R
\name{MAPE}
\alias{MAPE}
\title{Mean Absolute Percentage Error}
\usage{
MAPE(Actual, Fitted)
}
\arguments{
\item{Actual}{vector of "Actual" data}

\item{Fitted}{vector of "Fitted" data}
}
\description{
Mean Absolute Percentage Error.
}
\keyword{MAPE}
\keyword{error}
