% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlyToWeekly.R
\name{monthlyToWeekly}
\alias{monthlyToWeekly}
\title{Converts monthly data to weekly.}
\usage{
monthlyToWeekly(x)
}
\arguments{
\item{x}{data.table. One columns called "Month" (a Date), and variable(s) to convert.}
}
\description{
Roll the end of months based on days overlapping in the week.
}
\examples{
x <- data.table::data.table(Month = seq.Date(as.Date("2017-01-01"), as.Date("2017-12-01"), by ="month"))
x[, Var1 := c(sample(50:100, size = 12))]
x[, Var2 := c(sample(50:100, size = 12))]
x[, Var3 := c(sample(50:100, size = 12))]

x
monthlyToWeekly(x)
}
\keyword{monthly}
\keyword{weekly}
