% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namesAsFormula.R
\name{namesAsFormula}
\alias{namesAsFormula}
\title{Columns names to a formula layout.}
\usage{
namesAsFormula(x)
}
\arguments{
\item{x}{data.table to convert to formula}
}
\description{
Outputs columns of a t
}
\keyword{formula}
\keyword{lm}
\keyword{names}

