% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneDividedbyCoef.R
\name{oneDividedbyCoef}
\alias{oneDividedbyCoef}
\title{1/coef for a lm.}
\usage{
oneDividedbyCoef(lm.object)
}
\arguments{
\item{lm.object}{Output of lm function.}
}
\description{
How many of independent var needed to get 1 dependent var?
}
\keyword{coefficient}
\keyword{lm}

