% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/replaceNAWith0.R
\name{replaceNAWith0}
\alias{replaceNAWith0}
\title{Replaces all NAs with 0}
\usage{
replaceNAWith0(x)
}
\arguments{
\item{x}{A data.table}
}
\description{
Replaces all NAs with 0
}
\examples{
replaceNAWith0()
}
\keyword{NA}

