% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/replaceNAWith0inCertainCols.R
\name{replaceNAWith0inCertainCols}
\alias{replaceNAWith0inCertainCols}
\title{Replaces NAs with 0 in certain columns}
\usage{
replaceNAWith0inCertainCols(x, string)
}
\arguments{
\item{x}{A data.table}

\item{string}{A character string that the column name has}
}
\description{
Replaces all NAs with 0 in a column that has "string" in its name
}
\examples{
replaceNAWith0()
}
\keyword{NA}

