% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceNAWithinCertainCols.R
\name{replaceNAWithinCertainCols}
\alias{replaceNAWithinCertainCols}
\title{Replaces NAs in certain columns}
\usage{
replaceNAWithinCertainCols(x, replacement, string)
}
\arguments{
\item{x}{data.table.}

\item{string}{Character. A string that is in the target column name.}
}
\description{
Replaces all NAs with a specified replacement in a column that has "string" in its name
}
\examples{
x <- data.table::fread(system.file("extdata", "NA.csv", package = "nladwa"))
print(x)
replaceNAWithinCertainCols(x = x, replacement = "space", string = ".a")
print(x)
}
\keyword{NA}
