% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.ggplot.R
\name{resid.ggplot}
\alias{resid.ggplot}
\title{Plot residuals of a lm over time.}
\usage{
\method{resid}{ggplot}(lm.object, date.index)
}
\arguments{
\item{lm.object}{Output of lm function.}

\item{date.index}{Character. The column from the model source data that is the date variable.}
}
\description{
Prints a ggplot of residuals from a model over time.
}
\keyword{plot}
\keyword{residuals}

