% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residPlot.R
\name{residPlot}
\alias{residPlot}
\title{Plot residuals of a lm over time.}
\usage{
residPlot(lm.object, date.index)
}
\arguments{
\item{lm.object}{Output of lm function.}

\item{date.index}{Character. The column from the model source data that is the date variable.}
}
\description{
Prints a ggplot of residuals from a model over time.
}
\keyword{plot}
\keyword{residuals}

