% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Scatter plot with linear regression.}
\usage{
scatterPlot(data, x, y, title = NULL, ...)
}
\arguments{
\item{data}{data.table.}

\item{x}{Character. The column in data to be the x variable.}

\item{y}{Character. The column in data to be the y variable.}
}
\description{
Plots a scatter, with linear fit, with equation and r2 in the title.
}
\examples{
scatterPlot(mtcars, x = "wt", y = "mpg", title = "Cars")
}
\keyword{regression}
\keyword{scatter}

