% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumCertainCols.R
\name{sumCertainCols}
\alias{sumCertainCols}
\title{Sum certain columns}
\usage{
sumCertainCols(x, id.col = "Week", search.string, onlyTotal)
}
\arguments{
\item{x}{data.table.}

\item{id.col}{Character. ID column to index the output.}

\item{search.string}{Character. A string that is in the target column name.}

\item{onlyTotal}{Logical. TRUE = vector of summed columns, FALSE = column appended and returned with original data.table}
}
\description{
Replaces all NAs with a specified replacement in a column that has "string" in its name
}
\examples{
sumCertainCols(data.table::data.table(iris), id.col = "Species", search.string = "Petal", onlyTotal = FALSE)

sumCertainCols(data.table::data.table(iris), id.col = "Species", search.string = "Petal", onlyTotal = TRUE)
}
\keyword{columns}
\keyword{sum}
