% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainAndTest.sample.R
\name{TrainAndTest.sample}
\alias{TrainAndTest.sample}
\title{Create a data.table for actual v fitted, with train and test sets}
\usage{
TrainAndTest.sample(data, train.size = 0.75, date.index)
}
\arguments{
\item{data}{data.table.}

\item{train.size}{Proportion of dataset that should be in the training set. Between 0 and 1.}

\item{date.index}{Character. The date index of the lm(data = ?) variable.}
}
\description{
Takes a lm object and outputs an actual vs fitted table.
}
\keyword{actual}
\keyword{fitted}
\keyword{lm}
\keyword{sample}
\keyword{test}
\keyword{train}
