% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueCharCols.R
\name{uniqueCharCols}
\alias{uniqueCharCols}
\title{Prints the unique values of any character columns}
\usage{
uniqueCharCols(x)
}
\arguments{
\item{x}{data.table.}
}
\description{
Prints the unique values of any character columns
}
\examples{
x <- data.table::data.table(iris)
iris$Species <- as.character(iris$Species)
print(x)
uniqueCharCols(x)
}
\keyword{character}
\keyword{unique}

