% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsPlot.R
\name{varsPlot}
\alias{varsPlot}
\title{Plot multiple variables over time.}
\usage{
varsPlot(x, date.index, vars)
}
\arguments{
\item{x}{data.table}

\item{date.index}{Character. The column in x that is the date variable.}

\item{vars}{Vector of characters - columns from x to be plotted.}
}
\description{
Prints a ggplot of selected variables over time.
}
\keyword{plot}
\keyword{variables}

