% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wideToLong.R
\name{wideToLong}
\alias{wideToLong}
\title{Converts data from wide to long}
\usage{
wideToLong(x, id.vars, writeToClipboard)
}
\arguments{
\item{x}{data.frame or data.table. In wide format.}

\item{id.vars}{Character. Column name from the source data that will be the id of the output.}

\item{writeToClipboard}{Logical. TRUE or FALSE.}
}
\description{
Reads data from clipboard, and melts the data.frame based on the id.vars entered.
}
\examples{
wideToLong()
}
\keyword{long}
\keyword{reshape}
\keyword{wide}

