% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeModelToCSV.R
\name{writeModelToCSV}
\alias{writeModelToCSV}
\title{Linear model summary, actual v fitted, and decomp to CSV}
\usage{
writeModelToCSV(lm.object, filename, date.index)
}
\arguments{
\item{lm.object}{lm object. A lm saved to a variable.}

\item{filename}{Character string. What the output should be saved as.}

\item{date.index}{Character. The date index of the lm(data = ?) variable.}
}
\description{
Writes summary(lm.object), AvF, decomp to CSVs.
}
\keyword{AvF}
\keyword{CSV}
\keyword{decomp}
\keyword{lm}
\keyword{summary}

