% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/writeToClipboard.R
\name{writeToClipboard}
\alias{writeToClipboard}
\title{Writes a data.table or data.frame to the clipboard}
\usage{
writeToClipboard(x)
}
\arguments{
\item{x}{A data.table or data.frame}
}
\description{
Writes a data.table or data.frame to the clipboard.
}
\examples{
writeToClipboard()
}
\keyword{clipboard}
\keyword{write}

