% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/writeToCSV.R
\name{writeToCSV}
\alias{writeToCSV}
\title{Writes a data.table or data.frame to a CSV}
\usage{
writeToCSV(x, filename)
}
\arguments{
\item{x}{A data.table or data.frame}

\item{filename}{A character string that the output will be called, e.g. "output"}
}
\description{
Writes a data.table or data.frame to a CSV, with a filename manually entered.
}
\examples{
writeToCSV()
}
\keyword{CSV}
\keyword{write}

