% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/excelDateToRDate.R
\name{excelDateToRDate}
\alias{excelDateToRDate}
\title{Excel date to R date class}
\usage{
excelDateToRDate(Date)
}
\arguments{
\item{Date}{Date, as character in dd/mm/yyyy}
}
\description{
dd/mm/yyyy to yyyy-mm-dd
}
\examples{
excelDateToRDate()
}
\keyword{date}
\keyword{excel}

