% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCharAcrossCols.R
\name{findCharAcrossCols}
\alias{findCharAcrossCols}
\title{Find characters across all columns}
\usage{
findCharAcrossCols(x, string)
}
\arguments{
\item{x}{A data.table or data.frame}

\item{string}{Character. String that you want to find in x.}
}
\description{
Searches each column for a particular string
}
\examples{
x <- iris
findCharAcrossCols(x, "virg")
x[findCharAcrossCols(x, "virg")$Species, ]
}
\keyword{character}
\keyword{unique}
