% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixNames.R
\name{fixNames}
\alias{fixNames}
\title{Fixes column names}
\usage{
fixNames(x)
}
\arguments{
\item{x}{data.table. With weird column names.}
}
\description{
Replaces and weird characters or spaces with dots
}
\examples{
x <- data.table::fread(system.file("extdata", "Newspapers.csv", package = "nladwa"))
names(x)
fixNames(x)
names(x)
}
\keyword{names}
