% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.contribution.ggplot.R
\name{gam.contribution.ggplot}
\alias{gam.contribution.ggplot}
\title{Plots the contribution of a GAM's smoothed independent variables}
\usage{
gam.contribution.ggplot(gam.contribution.object)
}
\arguments{
\item{gam.contibution.object}{The output of gam.contribution().}
}
\description{
Takes gam.contribution output and visualises it.
}
\examples{
# see example from gam.contribution
}
\keyword{contribution}
\keyword{gam}
\keyword{ggplot}

