% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeek.R
\name{getWeek}
\alias{getWeek}
\title{Gets week from a date}
\usage{
getWeek(Date, order = "Mon-Sun")
}
\arguments{
\item{Date}{Date class object.}

\item{order}{Character. "Mon-Sun" or "Sun-Mon".}
}
\description{
Gets week from a date, has to be Date class
}
\examples{
x <- as.Date("2015-01-12")
getWeek(x, order = "Mon-Sun")
getWeek(x, order = "Sun-Mon")
}
\keyword{week}

