% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.contribution.R
\name{lm.contribution}
\alias{lm.contribution}
\title{Outputs the contribution of each lm independent variable}
\usage{
lm.contribution(lm.object, date.index = "Week")
}
\arguments{
\item{lm.object}{lm object. A lm saved to a variable.}

\item{date.index}{Character. The date index of the lm(data = ?) variable.}
}
\description{
Takes a lm and applies coefficients to show how each independent variable contributes to the dependent variable. IMPORTANT: lm(data = ?) has to be a data.table.
}
\examples{
x <- data.table::fread(system.file("extdata", "lm.csv", package = "nladwa"))
x[, Week := excelDateToRDate(Week)]
lm.x <- lm(TotalGrossLeads ~ Search.Spend + Competitor.Spend, data = x)
lm.object <- lm.contribution(lm.object = lm.x, date.index = "Week")
print(lm.object)
}
\keyword{contribution}
\keyword{lm}
