% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readExcelClipboard.R
\name{readExcelClipboard}
\alias{readExcelClipboard}
\title{Reads Excel table from clipboard to data.frame}
\usage{
readExcelClipboard()
}
\arguments{
\item{No}{arguments}
}
\description{
Create data.frame from Excel table that's held in the clipboard.
}
\examples{
readExcelClipboard()
}
\keyword{excel}

