% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/replaceNAWith.R
\name{replaceNAWith}
\alias{replaceNAWith}
\title{Replaces all NAs}
\usage{
replaceNAWith(x, replacement)
}
\arguments{
\item{x}{data.table.}
}
\description{
Replaces all NAs with whatever is specified.
}
\examples{
replaceNAWith()
}
\keyword{NA}

