% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uniqueCharCols.R
\name{uniqueCharCols}
\alias{uniqueCharCols}
\title{Prints the unique values of any character columns}
\usage{
uniqueCharCols(x)
}
\arguments{
\item{x}{data.table.}
}
\description{
Prints the unique values of any character columns
}
\examples{
uniqueCharCols()
}
\keyword{character}
\keyword{unique}

