% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsPlot.R
\name{varsPlot}
\alias{varsPlot}
\title{Plot multiple variables over time.}
\usage{
varsPlot(x, vars, date.index = "Week", facet = FALSE)
}
\arguments{
\item{x}{data.table}

\item{vars}{Vector of characters - columns from x to be plotted. If left blank function plots all variables.}

\item{date.index}{Character. The column in x that is the date variable.}

\item{facet}{TRUE = facets plot by variable, scales = "free_y".}
}
\description{
Prints a ggplot of selected variables over time.
}
\keyword{plot}
\keyword{variables}
