% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wideToLong.R
\name{wideToLong}
\alias{wideToLong}
\title{Coverts data from wide to long}
\usage{
wideToLong(id.vars)
}
\arguments{
\item{id.vars}{A vector, or individual column from the source data that will be the id of the output.}
}
\description{
Reads data from clipboard, and melts the data.frame based on the id.vars entered.
}
\examples{
wideToLong()
}
\keyword{long}
\keyword{reshape}
\keyword{wide}

